requirements.txt
Flask
opencv-contrib-python
numpy
pandas
Pillow
